move graphrag.index.cache to graphrag.cache
