+++ graphrag/api/index.py
@@ -10,9 +10,9 @@ Backwards compatibility is not guaranteed at this time.
 
 from pathlib import Path
 
+from graphrag.cache.noop_pipeline_cache import NoopPipelineCache
 from graphrag.config.enums import CacheType
 from graphrag.config.models.graph_rag_config import GraphRagConfig
-from graphrag.index.cache.noop_pipeline_cache import NoopPipelineCache
 from graphrag.index.create_pipeline_config import create_pipeline_config
 from graphrag.index.run import run_pipeline_with_config
 from graphrag.index.typing import PipelineRunResult
