+++ graphrag/cache/factory.py
@@ -16,14 +16,14 @@ from graphrag.storage.blob_pipeline_storage import BlobPipelineStorage
 from graphrag.storage.file_pipeline_storage import FilePipelineStorage
 
 if TYPE_CHECKING:
-    from graphrag.index.cache.pipeline_cache import PipelineCache
+    from graphrag.cache.pipeline_cache import PipelineCache
     from graphrag.index.config.cache import (
         PipelineCacheConfig,
     )
 
-from graphrag.index.cache.json_pipeline_cache import JsonPipelineCache
-from graphrag.index.cache.memory_pipeline_cache import InMemoryCache
-from graphrag.index.cache.noop_pipeline_cache import NoopPipelineCache
+from graphrag.cache.json_pipeline_cache import JsonPipelineCache
+from graphrag.cache.memory_pipeline_cache import InMemoryCache
+from graphrag.cache.noop_pipeline_cache import NoopPipelineCache
 
 
 def create_cache(
