+++ graphrag/cache/json_pipeline_cache.py
@@ -6,7 +6,7 @@
 import json
 from typing import Any
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.storage.pipeline_storage import PipelineStorage
