+++ graphrag/cache/noop_pipeline_cache.py
@@ -5,7 +5,7 @@
 
 from typing import Any
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 
 
 class NoopPipelineCache(PipelineCache):
