+++ graphrag/cache/pipeline_cache.py
