+++ graphrag/index/context.py
@@ -7,7 +7,7 @@
 from dataclasses import dataclass as dc_dataclass
 from dataclasses import field
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.storage.pipeline_storage import PipelineStorage
