+++ graphrag/index/llm/load_llm.py
@@ -27,7 +27,7 @@ from graphrag.llm import (
 if TYPE_CHECKING:
     from datashaper import VerbCallbacks
 
-    from graphrag.index.cache.pipeline_cache import PipelineCache
+    from graphrag.cache.pipeline_cache import PipelineCache
     from graphrag.index.typing import ErrorHandlerFn
 
 log = logging.getLogger(__name__)
