+++ graphrag/index/operations/embed_text/strategies/mock.py
@@ -9,7 +9,7 @@ from typing import Any
 
 from datashaper import ProgressTicker, VerbCallbacks, progress_ticker
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.operations.embed_text.strategies.typing import TextEmbeddingResult
