+++ graphrag/index/operations/embed_text/strategies/openai.py
@@ -11,7 +11,7 @@ import numpy as np
 from datashaper import ProgressTicker, VerbCallbacks, progress_ticker
 
 import graphrag.config.defaults as defs
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.llm.load_llm import load_llm_embeddings
 from graphrag.index.operations.embed_text.strategies.typing import TextEmbeddingResult
 from graphrag.index.text_splitting.text_splitting import TokenTextSplitter
