+++ graphrag/index/operations/embed_text/strategies/typing.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 
 from datashaper import VerbCallbacks
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 
 
 @dataclass
