+++ graphrag/index/operations/extract_covariates/extract_covariates.py
@@ -14,7 +14,7 @@ from datashaper import (
     derive_from_rows,
 )
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.operations.extract_covariates.typing import (
     Covariate,
     CovariateExtractStrategy,
