+++ graphrag/index/operations/extract_entities/strategies/graph_intelligence.py
@@ -6,7 +6,7 @@
 from datashaper import VerbCallbacks
 
 import graphrag.config.defaults as defs
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.graph.extractors import GraphExtractor
 from graphrag.index.llm.load_llm import load_llm
 from graphrag.index.operations.extract_entities.strategies.typing import (
