+++ graphrag/index/operations/summarize_communities/strategies.py
@@ -9,7 +9,7 @@ import traceback
 
 from datashaper import VerbCallbacks
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.graph.extractors.community_reports import (
     CommunityReportsExtractor,
 )
