+++ graphrag/index/operations/summarize_descriptions/strategies.py
@@ -5,7 +5,7 @@
 
 from datashaper import VerbCallbacks
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.graph.extractors.summarize import SummarizeExtractor
 from graphrag.index.llm.load_llm import load_llm
 from graphrag.index.operations.summarize_descriptions.typing import (
