+++ graphrag/index/update/incremental_index.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 from datashaper import VerbCallbacks
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.config.pipeline import PipelineConfig
 from graphrag.index.flows.generate_text_embeddings import generate_text_embeddings
 from graphrag.index.run.workflow import _find_workflow_config
