+++ graphrag/index/workflows/v1/subflows/create_base_entity_graph.py
@@ -14,7 +14,7 @@ from datashaper import (
 )
 from datashaper.table_store.types import VerbResult, create_verb_result
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.flows.create_base_entity_graph import (
     create_base_entity_graph as create_base_entity_graph_flow,
 )
