+++ graphrag/index/workflows/v1/subflows/create_final_covariates.py
@@ -13,7 +13,7 @@ from datashaper import (
 )
 from datashaper.table_store.types import VerbResult, create_verb_result
 
-from graphrag.index.cache.pipeline_cache import PipelineCache
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.flows.create_final_covariates import (
     create_final_covariates as create_final_covariates_flow,
 )
