+++ graphrag/storage/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""Storage package root."""
+"""The storage package root."""
