+++ graphrag/utils/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""Util functions for the GraphRAG run module."""
+"""Util functions for the GraphRAG package."""
