+++ graphrag/vector_stores/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""A module containing vector storage implementations."""
+"""A package containing vector store implementations."""
